requirements.txt
duckdb==1.5.1
numpy==2.4.4
pandas==3.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
